Add fast casting for classes that are final
Cleaning up a few static casts on the way.

Due to the new unified way to cast nodes, we can now
easily introduce a way to have a faster up-casting for
classes we know are final. This is done in a centralized
way. You simply keep calling `Cast` to up-cast to a more
specific type. The compiler will automatically choose the
faster type-id check, if the involving class is final. We
have to produce the list of the classes that act as bases
by hand. It turned out to be the much shorter list than
listing all final classes. Maybe the compiler could deduce
this information from a final keyword on the class, but his
approach seems safe and accomplishes the job either way.

FTR: Performance profiling showed that dynamic cast can
take up around 20% of the whole runtime. So having faster
casts for classes we now are safe seems to be a good thing.
Having its configuration centralized makes it also very
easy to maintain and regular code does not need to worry!