Speed up repeated extends by memoizing calculations
The extend code path is probably the most expensive to go
down in LibSass. It involves multiple conversions between
the regular AST selector node to the internal used format.

Storing results from previous runs is rather cheap, due to
the new shared object memory storage and since we already
have everything to use selectors efficiently as map keys.

It turned out that memoizing Compound-Selectors was too
much overhead. This is probably due to having a `Node`
object a map value (IMO this is a copy assignment). I
commented it out for now with some info for later re-
evaluation when we can store a more efficient value!